+++ tests/stm32/src/bin/stop.rs
@@ -10,9 +10,8 @@ use common::*;
 use cortex_m_rt::entry;
 use embassy_executor::Spawner;
 use embassy_stm32::Config;
-use embassy_stm32::low_power::{Executor, StopMode, stop_ready, stop_with_rtc};
+use embassy_stm32::low_power::{Executor, StopMode, reconfigure_rtc, stop_ready};
 use embassy_stm32::rcc::LsConfig;
-use embassy_stm32::rtc::{Rtc, RtcConfig};
 use embassy_time::Timer;
 
 #[entry]
@@ -57,7 +56,7 @@ async fn async_main(spawner: Spawner) {
         config.rcc.hsi = Some(HSIPrescaler::DIV4); // 64 MHz HSI will need a /4
     }
 
-    let p = init_with_config(config);
+    let _p = init_with_config(config);
     info!("Hello World!");
 
     let now = NaiveDate::from_ymd_opt(2020, 5, 15)
@@ -65,11 +64,7 @@ async fn async_main(spawner: Spawner) {
         .and_hms_opt(10, 30, 15)
         .unwrap();
 
-    let mut rtc = Rtc::new(p.RTC, RtcConfig::default());
-
-    rtc.set_datetime(now.into()).expect("datetime not set");
-
-    stop_with_rtc(rtc);
+    reconfigure_rtc(|rtc| rtc.set_datetime(now.into()).expect("datetime not set"));
 
     spawner.spawn(task_1().unwrap());
     spawner.spawn(task_2().unwrap());